farmerbot

+++ DigitalMeLib/servers/myjobs/MyJobs.py
@@ -1,5 +1,3 @@
-# from gevent import monkey
-# monkey.patch_all()
 
 import inspect
 from Jumpscale import j
@@ -64,24 +62,39 @@ class MyJobs(JSBASE):
         self.workers_nr_max = 10
         self.mainloop = None
         self.dataloop = None
+        self.workers_subprocess = True
         self.logger_enable()
 
-    def init(self,reset=False):
+    def init(self,reset=False,start=True):
         """
 
         :param db_cl: if None will use db_cl = j.clients.zdb.testdb_server_start_client_get(reset=True)
         :return:
         """
-        if self._init == False:
+        if self._init == False or reset:
 
-            if reset:
-                self.halt(reset=True)
+            if self.mainloop == None and  self.dataloop == None:
+                from gevent import monkey
+                monkey.patch_all()
+
+            if self.mainloop != None:
+                self.mainloop.kill()
+
+            if self.dataloop != None:
+                self.dataloop.kill()
 
             db = j.data.bcdb.get(j.core.db, namespace="myjobs", reset=reset,json_serialize=True)
 
             self.model_job = db.model_create(schema=schema_job)
             self.model_action = db.model_create(schema=schema_action)
             self.model_worker = db.model_create(schema=schema_worker)
+
+            if reset:
+                self.halt(reset=True)
+
+            if start:
+                self._start()
+
             self._init = True
 
 
@@ -102,21 +115,21 @@ class MyJobs(JSBASE):
     def nr_workers(self):
         return len(self.workers.values())
 
-    def start(self,onetime=False):
+    def _start(self,onetime=False):
         """
 
         :param nr_max: max nr of workers which can be started
         :return:
         """
-        self._worker_start(onetime=onetime) #start first worker
+        if self.workers_subprocess:
+            self._worker_start(onetime=onetime) #start first worker
         if onetime:
             self._data_process(onetime=True)
-            self._main_loop(onetime=True)
         else:
-            self.mainloop = gevent.spawn(self._main_loop)
             self.dataloop = gevent.spawn(self._data_process)
-            self.mainloop.start()
-            self.dataloop.start()
+            if self.workers_subprocess:
+                self.mainloop = gevent.spawn(self._main_loop)
+
 
     def _worker_start(self,onetime=False):
         w = self.model_worker.new()
@@ -129,6 +142,19 @@ class MyJobs(JSBASE):
         else:
             worker = gipc.start_process(target=myworker,args=(w.id,))
             self.workers[w.id] = worker
+
+    def worker_start_inprocess(self):
+        """
+        js_shell "j.servers.myjobs.worker_start_inprocess()"
+        :return:
+        """
+        self.init(reset=False,start=False)
+        w = self.model_worker.new()
+        w.time_start = j.data.time.epoch
+        w.last_update = j.data.time.epoch
+        w = self.model_worker.set(w)
+        self.logger.debug("worker started:%s"%w.id)
+        myworker(w.id,showout=True)
 
     def _data_process(self,onetime=False):
         while True:
@@ -155,12 +181,13 @@ class MyJobs(JSBASE):
                     raise RuntimeError("could not process returning data")
                 d = self.queue_data.get_nowait()
 
-            gevent.sleep(0.1) #important to not overload gedis
             if onetime:
                 return
 
+            gevent.sleep(1) #important to not overload gedis
 
-    def _main_loop(self,onetime=False):
+
+    def _main_loop(self):
 
         self.logger.debug("monitor start")
 
@@ -180,18 +207,16 @@ class MyJobs(JSBASE):
 
             self.logger.debug("monitor run")
 
-            #there is already 1 working, lets give 2 sec time before we start monitoring
-            time.sleep(2)
-
-            #test if we need to add workers
-            while test_workers_more():
-                print("WORKERS START")
-                self._worker_start()
+            # #there is already 1 working, lets give 2 sec time before we start monitoring
+            # time.sleep(2)
 
             #TEST for timeout
-            from pudb import set_trace; set_trace()
-
-            for wid, gproc in self.workers.items():
+            wids = [key for key in self.workers.keys()]
+            for wid in wids:
+                if wid in self.workers:
+                    gproc = self.workers[wid]
+                else:
+                    continue
                 if gproc.exitcode !=None:
                     raise RuntimeError("subprocess should never have been exitted")
                 w = self.model_worker.get(wid)
@@ -211,51 +236,63 @@ class MyJobs(JSBASE):
                         # print(w)
                         self.logger.info("KILL:%s in worker %s"%(w.id,job.id))
                         gproc.kill()
+                        self.workers.pop(wid)
                         job.state = "ERROR"
                         job.error = "TIMEOUT"
                         job.time_stop = j.data.time.epoch
                         self.model_job.set(job)
                         print(job)
-                        # restart worker
-                        self.workers[w.id] = gipc.start_process(target = self._worker,args=(w.id,))
+                        # make sure right nr of workers are active
+                        self._worker_start()
 
-            #test if we have too many workers
-            removed_one = False
-            active_workers = [key for key in self.workers.keys()]
-            active_workers.sort()
-            for wid in active_workers:
-                gproc = self.workers[wid]
-                if gproc.exitcode != None:
-                    raise RuntimeError("subprocess should never have been exitted")
-                w = self.model_worker.get(wid)
-                if w ==None:
-                    j.shell()
+            if test_workers_more():
+                #test if we need to add workers
+                while test_workers_more():
+                    print("WORKERS START")
+                    self._worker_start()
+            else:
 
-                job_running = w.current_job != 4294967295
-                self.logger.debug("job running:%s (%s)"%(job.id,job_running))
+                #test if we have too many workers
+                removed_one = False
+                active_workers = [key for key in self.workers.keys()]
+                active_workers.sort()
+                for wid in active_workers:
+                    gproc = self.workers[wid]
+                    if gproc.exitcode != None:
+                        raise RuntimeError("subprocess should never have been exitted")
+                    w = self.model_worker.get(wid)
+                    if w ==None:
+                        continue
 
-                if w.halt==False and not job_running and self.queue.qsize()==0:
-                    if removed_one == False and test_workers_less():
-                        self.logger.debug("worker remove:%s"%wid)
-                        removed_one = True
-                        w.halt = True
-                        self.model_worker.set(w) #mark worker to halt
-                        gproc.kill()
-                        self.model_worker.delete(wid)
-                        todel.append(wid)
+                    job_running = w.current_job != 4294967295
+                    self.logger.debug("job running:%s (%s)"%(w.id,job_running))
 
-            for item in todel:
-                self.workers.pop(item)
+                    if w.halt==False and not job_running and self.queue.qsize()==0:
+                        if removed_one == False and test_workers_less():
+                            self.logger.debug("worker remove:%s"%wid)
+                            removed_one = True
+                            w.halt = True
+                            self.model_worker.set(w) #mark worker to halt
+                            gproc.kill()
+                            gproc.terminate()
+                            self.model_worker.delete(wid)
+                            gproc2 = self.workers[wid]
+                            while gproc.is_alive():
+                                gevent.sleep(0.1)
+                                print("worker,killing:%s"%wid)
+                            assert gproc2.is_alive() == False
+                            self.workers.pop(wid)
 
-            print(self.workers)
+            # print(self.workers)
 
             self.logger.debug("nr workers:%s, queuesize:%s"%(self.nr_workers,self.queue.qsize()))
+            gevent.sleep(1)
 
-            if onetime:
-                return
 
 
     def schedule(self,method,*args,category="", timeout=120, inprocess=False, **kwargs):
+        if inprocess:
+            return method(*args,**kwargs)
         self.init()
         code = inspect.getsource(method)
         code = j.core.text.strip(code)
@@ -287,18 +324,21 @@ class MyJobs(JSBASE):
         job.kwargs = j.data.serializers.json.dumps(kwargs)
         job = self.model_job.set(job)
 
-        if inprocess:
-            j.shell()
-        else:
-            self.queue.put(job.id)
+        self.queue.put(job.id)
+
         return job.id
 
     def halt(self,graceful=True,reset=True):
-        self.init()
+
+        if self.mainloop != None:
+            self.mainloop.kill()
+
+        if self.dataloop != None:
+            self.dataloop.kill()
 
         for wid, gproc in self.workers.items():
             if gproc.exitcode != None:
-                raise RuntimeError("subprocess should never have been exitted")
+                continue
 
             w = self.model_worker.get(wid)
 
@@ -306,6 +346,7 @@ class MyJobs(JSBASE):
 
             if not graceful or not job_running:
                 gproc.kill()
+                gproc.terminate() #dont know difference
 
         if reset:
             self.model_action.destroy()
@@ -315,6 +356,26 @@ class MyJobs(JSBASE):
             while self.queue.get_nowait() != None:
                 pass
 
+    def results(self,ids,timeout=10):
+        res = {}
+        counter=0
+        while len(ids)>0:
+            id=ids[0]
+            job = self.model_job.get(id)
+            if job == None:
+                raise RuntimeError("job:%s not found"%id)
+            if job.time_stop!=0:
+                if job.state=="OK":
+                    res[id] = job.result
+                    ids=ids[1:]
+                else:
+                    raise RuntimeError("job in eror:\n%s"%job)
+            if len(ids) > 0:
+                gevent.sleep(0.1)
+                counter+=1
+                if counter>timeout*10:
+                    raise RuntimeError("timeout for results with jobids:%s"%ids)
+        return res
 
 
     def test(self):
@@ -322,7 +383,18 @@ class MyJobs(JSBASE):
         js_shell "j.servers.myjobs.test()"
         :return:
         """
-        self.init(reset=True)
+
+        def kill():
+            #kill leftovers from last time, if any
+            session = j.tools.tmux.session_get('main')
+            session.window_remove("myworker_worker")
+            self.init(reset=True,start=False)
+            jobs = self.model_job.get_all()
+            assert len(jobs)==0
+            assert self.queue.qsize() == 0
+            self.workers_subprocess = True
+
+        kill()
 
         def add(a,b):
             return a+b
@@ -339,25 +411,107 @@ class MyJobs(JSBASE):
             import time
             time.sleep(2)
 
+        #test the behaviour for 1 job in process, only gevent for data handling
+        j.servers.myjobs.schedule(add_error, 1, 2)
+        self._start(onetime=True)
+
+        jobs = self.model_job.get_all()
+        assert len(jobs) == 1
+        job = jobs[0]
+        assert job.error == "s"
+        assert job.id == 0
+        assert job.result == ""
+        assert job.state == "ERROR"
+        assert job.time_stop >0
+
+        j.servers.myjobs.schedule(add, 1, 2)
+        self._start(onetime=True)
+
+        jobs = self.model_job.get_all()
+        assert len(jobs) == 2
+        job = jobs[1]
+        assert job.error == ""
+        assert job.id == 1
+        assert job.result == '3'
+        assert job.state == "OK"
+        assert job.time_stop >0
+
+        res = self.results([1])
+
+        assert res == {1: '3'}  #is the result back
+
+        assert 3==j.servers.myjobs.schedule(add, 1, 2,inprocess=True)
+
+        #lets start from scratch, now we know the super basic stuff is working
+
+        self.workers_subprocess = False #will test with independent worker
+
+        self.init(reset=True)
+
         for x in range(10):
             self.schedule(add,1,2)
 
         j.servers.myjobs.schedule(add_error, 1, 2)
 
-        self.schedule(wait, timeout=1)
-
         jobs = self.model_job.get_all()
 
-        assert len(jobs) == 12
+        assert len(jobs) == 11
 
-        self.start()
+        assert self.queue.qsize() == 11 #there need to be 12 jobs in queue
 
-        time.sleep(3)
-        #now timeout should have happened
+        def tmuxexec():
+            #lets now run the job executor in tmux, see it runs well in process
+            cmd = 'js_shell "j.servers.myjobs.worker_start_inprocess()"'
+            j.tools.tmux.execute(
+                cmd,
+                session='main',
+                window='myworker_worker',
+                pane='main',
+                session_reset=False,
+                window_reset=True
+            )
+
+            time.sleep(1)
+            assert self.queue.qsize() == 0
+
+        tmuxexec()
+
+        print("WORKER STARTED IN TMUX")
+
+        print("will wait for results")
+        assert self.results([1, 2, 3],timeout=1) == {1: '3', 2: '3', 3: '3'}
 
         jobs = self.model_job.get_all()
+        errors = [job for job in jobs if job.state == "ERROR"]
+        assert len(errors) == 1
 
-        assert len(jobs) == 12
+        kill()
+
+        #TMUX and in process tests are done, lets now see if as subprocess it works
+
+        self.init(reset=True,start=True)
+
+        print("wait to schedule jobs")
+        gevent.sleep(2)
+
+        for x in range(20):
+            self.schedule(wait_2sec)
+
+        j.servers.myjobs.schedule(wait,timeout=1)
+        j.servers.myjobs.schedule(add_error, 1, 2)
+
+        print("there should be 10 workers, so wait is max 5 sec")
+        gevent.sleep(5)
+
+        #now timeout should have happened & all should have executed
+
+        jobs = self.model_job.get_all()
+        assert len(jobs) == 22
+
+        completed = [job for job in jobs if job.time_stop !=0]
+
+        assert len(completed)==22
+
 
         errors = [job for job in jobs if job.error != ""]
         assert len(errors) == 2
@@ -372,23 +526,12 @@ class MyJobs(JSBASE):
         assert len(errors) == 1
 
         jobs = [job for job in jobs if job.state == "OK"]
-        assert len(jobs) == 10
-        assert jobs[0].result == "3"
+        assert len(jobs) == 20
 
-        for x in range(50):
-            self.schedule(wait_2sec)
-
-
-        while len(self.model_worker.get_all())>1:
-            time.sleep(1)
-
-        j.shell()
+        self.halt(reset=True)
 
         print ("TEST OK")
 
-
-
-        self.halt(reset=True)
 
 
     def test2(self):
@@ -416,3 +559,23 @@ class MyJobs(JSBASE):
 
 
         self.halt(reset=True)
+
+    def test2_load(self):
+        """
+        js_shell "j.servers.myjobs.test2_load()"
+        :return:
+        """
+        self.init(reset=True)
+
+
+        def wait_2sec():
+            import time
+            time.sleep(2)
+            return "OK"
+
+        ids=[]
+        for x in range(40):
+            ids.append(self.schedule(wait_2sec))
+
+        res = self.results(ids)
+
